Shuffle around where helper functions and classes live so they can be called separately from matrix.py and metamath.py. The plan is to move these towards independent paths that users can switch between for v1.0
